Improve test definitions
