--- definitions/npm/countries-and-timezones_v1.x.x/test_countries-and-timezones_v1.x.x.js
@@ -1,21 +1,0 @@
-// @flow
-
-import timezones from 'countries-and-timezones';
-
-const allTimezones = timezones.getAllTimezones();
-const allCountries = timezones.getAllCountries();
-const timezone = timezones.getTimezonesForCountry('ES');
-const country = timezones.getCountriesForTimezone('Europe/Berlin');
-
-// $ExpectError
-(timezones.getAllTimezones(): string);
-// $ExpectError
-(timezones.getAllCountries(): string);
-// $ExpectError
-(timezones.getTimezonesForCountry('ES'): number);
-// $ExpectError
-timezones.getTimezonesForCountry();
-// $ExpectError
-(timezones.getCountriesForTimezone('Europe/Zurich'): void);
-// $ExpectError
-timezones.getCountriesForTimezone();
